recursos y usarios internos

+++ app/Http/Controllers/CicloController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Helpers\Debug;
 use App\Models\Ciclo;
 use App\Models\Curriculum;
 use App\Models\Requisito;
@@ -16,7 +17,8 @@ class CicloController extends Controller {
      *
      */
     public function index() {
-        $ciclos = Ciclo::whereHas('curriculum', function ($query) {$query->activo();})->WithRequisitos()->get();
+        $ciclos = Ciclo::whereHas('curriculum', function ($query) {$query->activo();})
+            ->WithCurriculum()->with('requisitos.cicloPre.curriculum')->get();
         return Inertia::render('Ciclos/index', [
             'ciclos' => $ciclos,
         ]);
@@ -46,21 +48,15 @@ class CicloController extends Controller {
             DB::beginTransaction();
             $ciclo = Ciclo::create($input);
             if ($ciclo) {
+                // foreach ($requisitos as $requi) {
+                //     $requi["ciclo_id"] = $ciclo->id;
+                //     Requisito::create($requi);
+                // }
+                $requisitosData = [];
                 foreach ($requisitos as $requi) {
-                    // $requiObj = (object) $requi;
-                    $requi["ciclo_id"] = $ciclo->id;
-                    // if (!empty($requiObj->delete)) {
-                    //     $item = Requisito::whereId($requiObj->id)->doesntHave('adicionalesInscripciones')->first();
-                    //     if (is_null($item)) {
-                    //         throw new Exception('Adicional con Inscripciones');
-                    //     }
-                    //     $item->delete();
-                    // } else if (!empty($requiObj->id)) {
-                    //     Requisito::find($requiObj->id)->update($requi);
-                    // } else {
-                    Requisito::create($requi);
-                    // }
+                    $requisitosData[] = array_merge($requi, ['ciclo_id' => $ciclo->id]);
                 }
+                $ciclo->requisitos()->createMany($requisitosData);
                 DB::commit();
                 return response()->json(["message" => "El Ciclo fue creada exitosamente!"], 200);
             } else {
@@ -81,7 +77,18 @@ class CicloController extends Controller {
      * @param  int  $id
      */
     public function show($id) {
-        $ciclo = Ciclo::whereId($id)->WithRequisitos()->with('requisitos.cicloPre.curriculum')->first();
+        $ciclo = Ciclo::whereId($id)
+            ->WithCurriculum()->WithRequisitos()
+            ->with('requisitos.cicloPre.curriculum.ciclos')->first();
+        $requisitos = $ciclo->requisitos->transform(function ($requi) {
+            return (object) [
+                'id' => $requi->id,
+                'ciclo_id' => $requi->ciclo_id,
+                'ciclo_pre_id' => $requi->ciclo_pre_id,
+                'curriculum' => $requi->cicloPre->curriculum,
+            ];
+        });
+        $ciclo->requisitos = $requisitos;
         $curriculums = Curriculum::activo()->with('ciclos')->get();
         return Inertia::render('Ciclos/form', [
             'action' => 'show',
@@ -96,7 +103,18 @@ class CicloController extends Controller {
      * @param  int  $id
      */
     public function edit($id) {
-        $ciclo = Ciclo::whereId($id)->WithRequisitos()->with('requisitos.cicloPre.curriculum')->first();
+        $ciclo = Ciclo::whereId($id)
+            ->WithCurriculum()->WithRequisitos()
+            ->with('requisitos.cicloPre.curriculum.ciclos')->first();
+        $requisitos = $ciclo->requisitos->transform(function ($requi) {
+            return (object) [
+                'id' => $requi->id,
+                'ciclo_id' => $requi->ciclo_id,
+                'ciclo_pre_id' => $requi->ciclo_pre_id,
+                'curriculum' => $requi->cicloPre->curriculum,
+            ];
+        });
+        $ciclo->requisitos = $requisitos;
         $curriculums = Curriculum::activo()->with('ciclos')->get();
 
         return Inertia::render('Ciclos/form', [
@@ -119,8 +137,8 @@ class CicloController extends Controller {
         $requisitos = $request->requisitos;
         try {
             DB::beginTransaction();
-            $ciclo = $ciclo->update($input);
-            if ($ciclo) {
+            $state = $ciclo->update($input);
+            if ($state) {
                 foreach ($requisitos as $requi) {
                     $requiObj = (object) $requi;
                     $requi["ciclo_id"] = $ciclo->id;
@@ -143,6 +161,7 @@ class CicloController extends Controller {
             }
 
         } catch (\Throwable $th) {
+            Debug::info($th);
             DB::rollBack();
             return response()->json(["message" => $th->getMessage(), 'server' => '¡El Ciclo no pudo ser actualizada, intente más tarde!'], 500);
 
@@ -156,8 +175,12 @@ class CicloController extends Controller {
      * @param  int  $id
      */
     public function destroy($id) {
-        $ciclo = Ciclo::whereId($id)->WithRequisitos()->first();
+        $ciclo = Ciclo::find($id);
+        if ($ciclo->dependientes->count() != 0) {
+            return response()->json(["message" => "", 'server' => '¡El Ciclo no puede ser elimiando porque depende de otro!'], 400);
+        }
         try {
+            $ciclo->requisitos()->delete();
             $state = $ciclo->delete();
             if ($state) {
                 return response()->json(["message" => "El Ciclo fue eliminada exitosamente!"], 200);
@@ -168,6 +191,5 @@ class CicloController extends Controller {
         } catch (\Throwable $th) {
             return response()->json(["message" => $th->getMessage(), 'server' => '¡El Ciclo no pudo ser eliminada, intente más tarde!'], 500);
         }
-
     }
 }
